requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
transformers==4.31.0
torch==2.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pillow==10.0.0
numpy==1.24.3
langdetect==1.0.9
sentencepiece==0.1.99